Instance methods should not write to "static" fields
