+++ C2CRIBuildDir/projects/C2C-RI/src/TMDDV3Verification/src/tmddv3verificationold/utilities/inouts/MessageValidator.java
@@ -63,7 +63,7 @@ public class MessageValidator implements ErrorHandler {
     /** Default memory usage report (false). */
     protected static final boolean DEFAULT_MEMORY_USAGE = false;
     /** Parse Error Variable (false). */
-    protected static boolean Parse_Result = false;
+    protected boolean Parse_Result = false;
     //
     // Data
     //
