+++ C2CRIBuildDir/projects/C2C-RI/src/TMDDV3Verification/src/tmddv3verificationold/utilities/inouts/PopulateDesignDetailTable.java
@@ -6,6 +6,9 @@
 package tmddv3verificationold.utilities.inouts;
 
 import org.apache.xmlbeans.SchemaType;
+import org.apache.xmlbeans.XmlCursor;
+import org.apache.xmlbeans.XmlObject;
+import org.apache.xmlbeans.XmlOptions;
 import tmddv3verificationold.utilities.TMDDDatabase;
 import tmddv3verificationold.utilities.TMDDWSDL;
 
@@ -28,10 +31,33 @@ public class PopulateDesignDetailTable {
 
         SchemaType[] globalElems = thisWSDL.getSchemaTypeSystem().documentTypes();
         for (int ii = 0; ii < globalElems.length; ii++) {
-             String result = SampleXmlUtilDoctored.createSampleForType(globalElems[ii]);
+             createSampleForType(globalElems[ii]);
         }
         SampleXmlUtilDoctored.storeResults();
 
+    }
+	
+	    public static String createSampleForType(SchemaType sType) {
+  
+        XmlObject object = XmlObject.Factory.newInstance();
+        XmlCursor cursor = object.newCursor();
+        // Skip the document node
+        cursor.toNextToken();
+        // Using the type and the cursor, call the utility method to get a
+        // sample XML payload for that Schema element
+		SampleXmlUtilDoctored oSample = new SampleXmlUtilDoctored(false);
+		SampleXmlUtilDoctored.messageName = sType.getDocumentElementName().getLocalPart();
+		oSample.lastParentType = SampleXmlUtilDoctored.messageName;
+        oSample.createSampleForType(sType, cursor);
+        // Cursor now contains the sample payload
+        // Pretty print the result.  Note that the cursor is positioned at the
+        // end of the doc so we use the original xml object that the cursor was
+        // created upon to do the xmlText() against.
+        XmlOptions options = new XmlOptions();
+        options.put(XmlOptions.SAVE_PRETTY_PRINT);
+        options.put(XmlOptions.SAVE_PRETTY_PRINT_INDENT, 2);
+        options.put(XmlOptions.SAVE_AGGRESSIVE_NAMESPACES);
+        return object.xmlText(options);
     }
 
     public static void clearTable() {
